light-engine/requirements.txt
PySide6
playwright
logging
beautifulsoup4
fastapi
uvicorn
pyinstaller
httpx
websockets
requests
python-dotenv
psutil